theme_common/requirements.txt
mkdocs-material>=9.5.2,<10.0.0
mkdocs-glightbox>=0.3.5,<1.0.0
pyyaml>=6.0.1,<7.0.0
